fix: LoadBotJoin opcode is u16 (0x0BB0), 24-byte body; positive spawn coords
Emu answered the bridge request: LoadBotJoin opcode = 0x0BB0 and it's a u16
GameOpcode (not u32), so the body is 24 bytes. Update the encoder, BotConfig.
JoinOpcode (uint->ushort), the test vectors, and the contract doc. Harness
parses the opcode as u16 and defaults to positive spawn coords + a strictly-
positive-in-X/Z region (the emu's zone bounds-check rejects X or Z <= 0).

+++ src/EqoaLoadClient.Core/Bot/BotConfig.cs
@@ -11,7 +11,7 @@ public sealed class BotConfig
     public byte ClassId { get; init; }
     public byte Level { get; init; }
     public ushort Cluster { get; init; }
-    public uint JoinOpcode { get; init; }          // from the emu registry
+    public ushort JoinOpcode { get; init; }        // u16 GameOpcode from the emu registry (0x0BB0)
     public int IntervalMs { get; init; } = 100;
     public required IMovementRegion Region { get; init; }
 }
